refactor(fiori-mcp-server): split api-json.ts by responsibility
api-json.ts mixed ui5.yaml/version discovery, disk cache + HTTP fetch, and
control/chain resolution under a name describing only one of them. Split
into ui5-config-resolver.ts (yaml discovery + version resolution) and
api-json-cache.ts (disk cache + fetch + fallback chain), leaving api-json.ts
for findControl/resolveLibraryForClass/resolveControlChain. Test files are
split to mirror the modules so the caching policy is testable on its own.
Pure code move — no behaviour change.

Addresses PR review feedback.
